requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.1
django-sortedm2m==1.5.0
django-taggit==1.1.0
docopt==0.6.2
ExifRead==2.1.2
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.0.0
pipreqs==0.4.9
pycodestyle==2.5.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
urllib3==1.24.3
yarg==0.1.9
psycopg2==2.9.3



